requirements.txt
mysql-connector-python==8.0.22
black==20.8b1
praw==7.1.4
